requirements.txt
gunicorn==20.0.4
click==7.1.2
Flask==1.1.2
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
Pillow==7.1.2
PyYAML==5.3.1
Werkzeug==1.0.1
https://download.pytorch.org/whl/cpu/torch-1.5.0%2Bcpu-cp38-cp38-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.6.0%2Bcpu-cp38-cp38-linux_x86_64.whl
